Rename tree files to match class name
