+++ src/explorer/AppSettingsTreeItem.ts
